Conflicts:
	Adroit.Accounting/Adroit.Accounting.SQL/Adroit.Accounting.SQL.csproj
	Adroit.Accounting/Adroit.Accounting.Web2/Program.cs

+++ Adroit.Accounting/Adroit.Accounting.Repository/IRepository/ISoftware.cs
@@ -11,7 +11,7 @@ namespace Adroit.Accounting.Repository.IRepository
         List<DropdownViewModel> SelectList(string connectionString);
         List<DropdownViewModel> SelectList(short businessId, string connectionString);
         List<SoftwareGridViewModel> List(string connectionString, int loginId = 0, int firmId = 0, string search = "", int pageStart = 0, int pageSize = 10, int sortColumn = 0, string sortOrder = "ASC");
-        int GetSoftwareIdByLoginId(int loginId, string connectionString);
+        byte GetSoftwareIdByLoginId(int loginId, string connectionString);
 
     }
 }
